demrepofdev/al5int/redefine_keyboard: Reworked keyboard control.  Key redefinition now working

+++ src/host_abstraction_layer/allegro_5/keyboard_internal.h
@@ -6,7 +6,7 @@
 #include "host_abstraction_layer/event_handler.h"
 
 extern bool keydefining;
-extern int keylookup[ALLEGRO_KEY_MAX];
+extern uint8_t keylookup[ALLEGRO_KEY_MAX];
 
 elk_event_t keyboard_handle_event(ALLEGRO_EVENT *event);
